Refactor: made changes to imports and displayName
