+++ src/modules/core/components/Button/ThreeDotsButton.tsx
@@ -7,7 +7,7 @@ import Icon from '~core/Icon';
 import Button, { Props as DefaultButtonProps } from './Button';
 import styles from './ThreeDotsButton.css';
 
-const displayName = 'ThreeDotsButton';
+const displayName = 'Button.ThreeDotsButton';
 
 interface Props extends DefaultButtonProps {
   activeStyles?: string;
