+++ src/modules/core/components/Button/index.ts
@@ -4,3 +4,4 @@ export { default as ConfirmButton } from './ConfirmButton';
 export { default as DottedAddButton } from './DottedAddButton';
 export { default as DialogActionButton } from './DialogActionButton';
 export { default as IconButton } from './IconButton';
+export { default as ThreeDotsButton } from './ThreeDotsButton';
